move Nav into top _layout.svelte file
